Paljudele päringutele lisatud kasutajapõhise juurdepääsupiirangu kontroll.

+++ src/main/java/ee/adit/ws/endpoint/document/DeleteDocumentEndpoint.java
@@ -112,6 +112,14 @@ public class DeleteDocumentEndpoint extends AbstractAditBaseEndpoint {
 				String errorMessage = this.getMessageSource().getMessage("user.inactive", new Object[] { userCode }, Locale.ENGLISH);
 				throw new AditException(errorMessage);
 			}
+			
+			// Check whether or not the application has rights to
+			// modify current user's data.
+			int applicationAccessLevelForUser = userService.getAccessLevelForUser(applicationName, user);
+			if(applicationAccessLevelForUser != 2) {
+				String errorMessage = this.getMessageSource().getMessage("application.insufficientPrivileges.forUser.write", new Object[] { applicationName, user.getUserCode() }, Locale.ENGLISH);
+				throw new AditException(errorMessage);
+			}
 				
 			this.getDocumentService().DeleteDocument(request.getDocumentId(), userCode, applicationName);
